refactor(bottomnav): move BottomNavOrder screen to impl, delegate via BlocScreen
BottomNavOrderScreen and BottomNavOrderPreviews move from
bottomnav/public to bottomnav/impl/ui. BottomNavOrderBloc extends
BlocScreen; BottomNavOrderBlocImpl overrides Content() to call
BottomNavOrderScreen(this). RootBloc.Child.BottomNavOrder gains
BlocScreen by bloc; RootScreen renders via child.Content().

`BottomNavBloc.Tab.getLabel()` / `.getIcon()` lose `internal` so
the moved BottomNavOrderScreen (now in bottomnav/impl) can reuse
them. They'll move into impl alongside BottomNavScreen when the
parent BottomNav is converted; visibility can tighten again then.

bottomnav/public sets publicResClass = true so the moved screens
access the generated Res accessors; bottomnav/impl gains the
Compose plugin + ui/text deps mirrored from public.
