+++ client/bottomnav/impl/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
     alias(libs.plugins.kotlinSerialization)
 }
 
@@ -7,9 +8,13 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.client.bottomnav.public)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(projects.client.shared)
             implementation(libs.kotlinx.serialization.json)
+            implementation(compose.components.resources)
             api(projects.client.browser.public)
             api(projects.client.grocery.core.public)
             api(projects.client.meal.core.public)
