+++ client/bottomnav/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/impl/ui/BottomNavOrderPreviews.kt
@@ -1,7 +1,11 @@
-package com.plusmobileapps.chefmate.recipe.bottomnav
+package com.plusmobileapps.chefmate.recipe.bottomnav.impl.ui
 
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc
+import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavOrderBloc
+import com.plusmobileapps.chefmate.recipe.bottomnav.DEFAULT_TAB_ORDER
 import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
 import kotlinx.coroutines.flow.MutableStateFlow
 
@@ -14,6 +18,11 @@ private fun bottomNavOrderBloc(model: BottomNavOrderBloc.Model): BottomNavOrderB
         override fun onSave() = Unit
 
         override fun onBack() = Unit
+
+        @Composable
+        override fun Content(modifier: Modifier) {
+            BottomNavOrderScreen(bloc = this, modifier = modifier)
+        }
     }
 
 val previewBottomNavOrderBloc: BottomNavOrderBloc = bottomNavOrderBloc(BottomNavOrderBloc.Model())
