+++ client/bottomnav/public/build.gradle.kts
@@ -21,4 +21,6 @@ kotlin {
     }
 }
 
+compose { resources { publicResClass = true } }
+
 plusLibrary { namespace = "com.plusmobileapps.chefmate.recipe.bottomnav" }
