+++ client/root/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/root/RootScreen.kt
@@ -31,7 +31,6 @@ import com.plusmobileapps.chefmate.browser.BrowserRootScreen
 import com.plusmobileapps.chefmate.devsettings.DeveloperSettingsScreen
 import com.plusmobileapps.chefmate.featureflag.FeatureFlagsScreen
 import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailScreen
-import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavOrderScreen
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavigationScreen
 import com.plusmobileapps.chefmate.recipe.core.addmeal.MealPlannerRootScreen
 import com.plusmobileapps.chefmate.recipe.core.root.RecipeRootScreen
@@ -157,7 +156,7 @@ private fun RootChildContent(child: RootBloc.Child, onBack: () -> Unit) {
             )
         is RootBloc.Child.MealPlanner -> MealPlannerRootScreen(child.bloc)
         is RootBloc.Child.AppSettings -> AppSettingsScreen(child.bloc)
-        is RootBloc.Child.BottomNavOrder -> BottomNavOrderScreen(child.bloc)
+        is RootBloc.Child.BottomNavOrder -> child.Content()
         is RootBloc.Child.DeveloperSettings -> DeveloperSettingsScreen(child.bloc)
         is RootBloc.Child.FeatureFlags -> FeatureFlagsScreen(child.bloc)
         is RootBloc.Child.CookMode -> child.Content()
